deal with XML when doing CJK extend on titles: https://github.com/metanorma/isodoc/issues/617
